Metric creation and deletion works
