+++ src/components/canvas-space/canvas-funct.jsx
