+++ src/components/canvas-space/canvas-guesstimate.jsx
