+++ src/components/canvas-space/canvas-space.styl
@@ -1,0 +1,44 @@
+.canvas-space
+  .metric
+    height 100%
+    background #ECECED
+    border-radius 2px
+    padding 0.5em
+    font-family: Lato-Light;
+    .row1
+      margin-bottom 0.25em
+    input
+      width 100%
+      border none
+      background #fff
+      border-radius 2px
+      padding 2px
+      &:focus
+        outline none
+        background #C1E1E6
+    &:focus
+      outline none
+    .name
+      font-size: 1.25em
+      color: #595F6F;
+    .median
+      font-size: 2em
+      color: #595F6F;
+    .nav > li > a
+      padding 4px 8px
+    .distribution
+      table
+        color: #666
+      h2
+        font-size: 25px;
+      h2.mean
+        color #3782c4
+        margin-bottom 0.25em
+        margin-top 0.25em
+
+      &:focus
+        background blue
+      h2.metric-name
+        margin-top 0
+        #contenteditable:focus
+          outline none
