+++ src/components/grid/grid.jsx
@@ -3,7 +3,7 @@
 import React from 'react'
 import _ from 'lodash'
 
-import styles from './cell.styl'
+import styles from './grid.styl'
 import Cell from './cell'
 import {keycodeToDirection, DirectionToLocation} from './utils'
 
@@ -18,40 +18,34 @@ const Grid = React.createClass({
       }
     };
   },
-  handleSelect(location) {
-    this.props.onMove(location)
-  },
   keyPress(e) {
     let direction = keycodeToDirection(e.keyCode)
     if (direction) {
       e.preventDefault()
       let newLocation = new DirectionToLocation(this.props.size, this.props.selected)[direction]()
-      this.handleSelect(newLocation)
+      this.props.handleSelect(newLocation)
     }
   },
-  cell(location) {
+  _cell(location) {
    let atThisLocation = (l) => _.isEqual(l, location)
    let isSelected = atThisLocation(this.props.selected)
-   if (isSelected) {
-   }
    let item = this.props.children.filter(function(i) { return (atThisLocation(i.props.item.location)) })[0];
    return (
-    <td>
+    <td key={'grid-item', location.row, location.column}>
       <Cell
-        key={'grid-item', location.row, location.column}
         location={location}
         isSelected={isSelected}
         item={item}
         gridKeyPress={this.keyPress}
-        onSelectItem={this.handleSelect}
+        handleSelect={this.props.handleSelect}
         onAddItem={this.props.onAddItem}/>
     </td>
     )
   },
-  row(row) {
+  _row(row) {
     return (
       upto(this.props.size.columns).map((column) => {
-        return(this.cell({row: row, column: column}))
+        return(this._cell({row: row, column: column}))
       })
     )
   },
@@ -61,7 +55,7 @@ const Grid = React.createClass({
         <table>
           {
             upto(this.props.size.rows).map((row) => {
-              return ( <tr> {this.row(row)} </tr>)
+              return ( <tr key={row}> {this._row(row)} </tr>)
             })
           }
         </table>
