+++ src/components/grid/grid.styl
@@ -7,6 +7,9 @@
     td
       padding 5px 10px;
       .cell
+        .cell-full
+          width 100%
+          height 100%j
         border-radius: 3px
         width: 100%;
         height: 90px
@@ -14,7 +17,5 @@
           outline none
         &.empty:focus
           background-color: #D6DCFF
-        &.full
-          background-color #A7ADD2
         &:hover
           background-color: #ddd;
