+++ src/css/main.styl
@@ -24,26 +24,7 @@ $icon-font-path = '../../node_modules/bootstrap-styl/fonts/'
   height: 100%
   border-radius: 3px
   background-color #aaa
-  
-.metric
-  .nav > li > a
-    padding 4px 8px
-  .distribution
-    table
-      color: #666
-    h2
-      font-size: 25px;
-    h2.mean
-      color #3782c4
-      margin-bottom 0.25em
-      margin-top 0.25em
 
-    &:focus
-      background blue
-    h2.metric-name
-      margin-top 0
-      #contenteditable:focus
-        outline none
 
 .refresh-calculation
   position absolute
